Restructure common package
